requirements.txt
streamlit
pytesseract
pillow
google-auth
google-auth-oauthlib
google-api-python-client